Add time precision functionality
Requirement is for time to be recorded in seconds. This commit
implements the necessary functionality and introduces corresponding
tests.

Note: this commit also removes a faulty test which will always fail
because of a change to the way the application works (widget_visible()
will always return false once it has set scroll_to_widget_event_recorded
as true.